Accept token and tenant_id auth
Allow swiftclient to authenticate using tenant name/id and token
but not storage url. This is useful when the client is used by
other services to access a users swift account. In such cases, the
password is may not be provided, but the service can discover the
storage endpoint by reauthenticating using the token and tenant
name/id.

Change-Id: I4477af445474c5fa97ff864c4942f1330b59e5d6